+++ printing/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add imageFromAssetBundle and networkImage
 - Add Page orientation on PdfPreview
 - Improve PrintJob object
+- Implement dynamic layout on iOS and macOS
 
 ## 5.0.0-nullsafety.1